Move everything metric test related to api/metric/metrictest
